// packages/x-core/src/nodes/enums.ts
export const enum NodeTypes {
  Undefined = "UNDEFINED",
  Null = "NULL",
  String = "STRING",
  Number = "NUMBER",
  Boolean = "BOOLEAN",
  Object = "OBJECT",
  Array = "ARRAY",
  Function = "FUNCTION",
  RegExp = "REGEXP",
  Symbol = "SYMBOL",
  BigInt = "BIGINT",
  Map = "MAP",
  Set = "SET",
  WeakMap = "WEAK_MAP",
  WeakSet = "WEAK_SET",
  JSValue = "JS_VALUE"
}